Revert "Will calendar"
