+++ components/SearchBar.js
@@ -7,16 +7,11 @@ import {
   Modal,
   TouchableOpacity
 } from "react-native";
-
 import { text } from "../constants/Styles";
 import apiKey from "../apiKey";
 import { GooglePlacesAutocomplete } from "react-native-google-places-autocomplete";
 import { Ionicons, MaterialCommunityIcons } from "@expo/vector-icons";
 import Colors from "../constants/Colors";
-
-import window from "../constants/Layout";
-import Calendar from "./Calendar";
-
 import { width, height } from "../constants/Layout";
 
 const homePlace = {
@@ -31,7 +26,6 @@ const workPlace = {
 const ASPECT_RATIO = width / height;
 const LATITUDE_DELTA = 0.0922;
 const LONGITUDE_DELTA = LATITUDE_DELTA * ASPECT_RATIO;
-
 
 class SearchBar extends React.Component {
   state = {
@@ -197,8 +191,8 @@ class SearchBar extends React.Component {
                 name={Platform.OS === "ios" ? "ios-close" : "md-close"}
               />
             </TouchableOpacity>
-            {/* <Text>Choisir une date</Text> */}
-            <Calendar />
+
+            <Text>Choisir une date</Text>
           </View>
         </Modal>
 
@@ -315,7 +309,7 @@ const styles = StyleSheet.create({
 
   containerDateModal: {
     backgroundColor: "white",
-    // padding: 22,
+    padding: 22,
     alignItems: "center",
     justifyContent: "flex-start",
     marginTop: 300,
