+++ constants/Layout.js
@@ -3,7 +3,7 @@ import { Dimensions } from "react-native";
 const width = Dimensions.get("window").width;
 const height = Dimensions.get("window").height;
 
-module.export = {
+module.exports = {
   width,
   height,
   isSmallDevice: width < 375
