+++ navigation/MainTabNavigator.js
@@ -13,6 +13,7 @@ import MyBikesScreen from "../screens/MyBikes";
 import AddBike from "../screens/AddBike";
 import AddPayementMethod from "../screens/AddPayementMethod";
 import AccountScreen from "../screens/AccountScreen";
+import Calendar from "../screens/Calendar";
 import EndRent from "../screens/EndRent";
 import Filters from "../screens/Filters";
 import List from "../screens/List";
@@ -28,6 +29,7 @@ import UserProfile from "../screens/UserProfile";
 const HomeStack = createStackNavigator({
   Home: { screen: HomeScreen, navigationOptions: { header: null } },
   Filters: { screen: Filters },
+  Calendar: { screen: Calendar, navigationOptions: { header: null } },
   List: { screen: List, navigationOptions: { header: null } },
   BikeDetails: { screen: BikeDetails },
   UserProfile: { screen: UserProfile },
