+++ package.json
@@ -14,10 +14,8 @@
     "@expo/samples": "2.1.1",
     "axios": "^0.18.0",
     "expo": "^31.0.2",
-    "moment": "^2.22.2",
     "react": "16.5.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-31.0.0.tar.gz",
-    "react-native-calendar-picker": "^5.22.0",
     "react-native-google-places-autocomplete": "^1.3.9",
     "react-navigation": "^2.18.2"
   },
